Make baseurl interpretation consistent
Follow each `{{ .Site.BaseURL }}` with a single `/`. This is consistent
with the interpretation in layouts/partials/header.html and fixes an
issue where users cannot host a hugo site from the root directory
without specifying the fqdn.

Now it is possible to do so by setting the baseurl to the empty string.
It would be more elegant to normalise BaseURL usage to expect a string
that ends in a slash, but that would break any existing setup where
BaseURL does not end in a slash.